Relationships with secondary model.
Added the SwiftUI searchable() functionality, which provides a complete and well-designed search bar with just a couple of lines of code. Added the sights attribute to the Destination Class Model, which necessitates the creation of a secondary class model named Sight. Implemented dynamic filtering based on user input to display Destination objects that match what the user has typed in the search bar. Finally, added a sights section to the EditDestinationView to display all the sights related to a specific destination, as well as a text field to add as many sights as desired, with the functionality to swipe to delete.